Move container generator volume specific things to volumes.go file
